feat: make doc command use pager
